CRUD + 페이징&검색 대부분 완성 - 20210709 18:57

+++ funding/src/main/java/org/thesix/funding/entity/Funding.java
@@ -33,7 +33,7 @@ public class Funding extends BaseEntity {
 
     private boolean removed;  // 펀딩 삭제여부
 
-    private int totalAmount; // 토탈 모금 금액
+    private long totalAmount; // 토탈 모금 금액
 
 
     /**
@@ -55,4 +55,5 @@ public class Funding extends BaseEntity {
     public void changeRemoved(boolean removed){
         this.removed = removed;
     }
+
 }
